+++ examples/aws-vault/devenv.nix
@@ -2,7 +2,7 @@
 
 {
   # Since Terraform adopted a non-free license (BSL 1.1) in August 2023,
-  # using terraform instead of opentofu now requires adding `allowUnfree: true` to `devenv.yaml`
+  # using terraform instead of opentofu now requires adding `allow_unfree: true` to `devenv.yaml`
   languages.opentofu.enable = true;
 
   aws-vault = {